fixed problems with last thread

+++ zhuravlev.ivan/I2/cnt_max_elem.cpp
@@ -3,7 +3,7 @@
 size_t counterMaxElement(const int *arr, const size_t size)
 {
   size_t cnt_max = 0;
-  size_t max = 0;
+  int max = 0;
   for (size_t i = 0; i < size; ++i)
   {
     if (arr[i] > max)
